Fix: Hooks pattern to refactor code
