+++ src/components/alert/SuccessAlert.jsx
@@ -1,0 +1,18 @@
+import * as React from 'react';
+import Snackbar from '@mui/material/Snackbar';
+import Alert from '@mui/material/Alert';
+
+export const SuccessAlert = ({open, handleClose, message}) => {
+  return (
+    <Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
+        <Alert
+          onClose={handleClose}
+          severity="success"
+          variant="filled"
+          sx={{ width: '100%' }}
+        >
+          {message}
+        </Alert>
+    </Snackbar>
+  );
+}
